+++ app/routes/_auth/auth.$provider/index.ts
@@ -4,7 +4,7 @@ import { handleMockAction } from '#app/utils/connections.server.ts'
 import { ProviderNameSchema } from '#app/utils/connections.tsx'
 import { getReferrerRoute } from '#app/utils/misc.tsx'
 import { getRedirectCookieHeader } from '#app/utils/redirect-cookie.server.ts'
-import { type Route } from './+types/auth_.$provider.ts'
+import { type Route } from './+types/index.ts'
 
 export async function loader() {
 	return redirect('/login')
