+++ app/routes/_auth/onboarding/$provider.server.ts
@@ -1,8 +1,8 @@
 import { invariant } from '@epic-web/invariant'
 import { redirect } from 'react-router'
 import { verifySessionStorage } from '#app/utils/verification.server.ts'
-import { onboardingEmailSessionKey } from './onboarding.tsx'
-import { type VerifyFunctionArgs } from './verify.server.ts'
+import { type VerifyFunctionArgs } from '../verify.server.ts'
+import { onboardingEmailSessionKey } from './index.tsx'
 
 export async function handleVerification({ submission }: VerifyFunctionArgs) {
 	invariant(
