+++ app/routes/_auth/onboarding/$provider.tsx
@@ -29,8 +29,8 @@ import { authSessionStorage } from '#app/utils/session.server.ts'
 import { redirectWithToast } from '#app/utils/toast.server.ts'
 import { NameSchema, UsernameSchema } from '#app/utils/user-validation.ts'
 import { verifySessionStorage } from '#app/utils/verification.server.ts'
-import { type Route } from './+types/onboarding_.$provider.ts'
-import { onboardingEmailSessionKey } from './onboarding'
+import { type Route } from './+types/$provider.ts'
+import { onboardingEmailSessionKey } from './index.tsx'
 
 export const providerIdKey = 'providerId'
 export const prefilledProfileKey = 'prefilledProfile'
