+++ app/routes/_auth/onboarding/index.tsx
@@ -24,7 +24,7 @@ import {
 	UsernameSchema,
 } from '#app/utils/user-validation.ts'
 import { verifySessionStorage } from '#app/utils/verification.server.ts'
-import { type Route } from './+types/onboarding.ts'
+import { type Route } from './+types/index.ts'
 
 export const onboardingEmailSessionKey = 'onboardingEmail'
