+++ app/routes/admin/cache/lru.$cacheKey.ts
@@ -1,12 +1,12 @@
 import { invariantResponse } from '@epic-web/invariant'
-import { cache } from '#app/utils/cache.server.ts'
+import { lruCache } from '#app/utils/cache.server.ts'
 import {
 	getAllInstances,
 	getInstanceInfo,
 	ensureInstance,
 } from '#app/utils/litefs.server.ts'
 import { requireUserWithRole } from '#app/utils/permissions.server.ts'
-import { type Route } from './+types/cache_.sqlite.$cacheKey.ts'
+import { type Route } from './+types/lru.$cacheKey.ts'
 
 export async function loader({ request, params }: Route.LoaderArgs) {
 	await requireUserWithRole(request, 'admin')
@@ -26,6 +26,6 @@ export async function loader({ request, params }: Route.LoaderArgs) {
 			isPrimary: currentInstanceInfo.primaryInstance === instance,
 		},
 		cacheKey,
-		value: cache.get(cacheKey),
+		value: lruCache.get(cacheKey),
 	}
 }
