+++ app/routes/admin/cache/sqlite.tsx
@@ -1,0 +1,1 @@
+export { action } from './sqlite.server.ts'
