--- app/routes/admin/cache_.sqlite.tsx
@@ -1,1 +1,0 @@
-export { action } from './cache_.sqlite.server.ts'
