+++ app/routes/users/$username/index.test.tsx
