+++ app/routes/users/$username/notes/+shared/note-editor.server.tsx
