+++ app/routes/users/$username/notes/new.tsx
